refactor Analyze class methods to base Strategy class
